MSDradix is broken, gotta fix some refactoring/fromatting changes everywhere else add my_sort2, O(n^2) alg that looks kinda cool, slow though

+++ compare/sdl_radix_compare.c
@@ -4,5 +4,5 @@
 #include "sdl_radix_compare.h"
 
 int sdl_radix_compare(void *a, void *b) {
-	return *(uint64_t *)a & (uint64_t)((SDL_Rect *)b)->h;
+	return *(uint32_t *)a & (uint32_t)((SDL_Rect *)b)->h;
 }
